divide tests to multiple files
