+++ src/tests/enter_exit/pseudo.rs
@@ -1,0 +1,3 @@
+mod enter_exit;
+
+pub use self::enter_exit::*;
